V1.1.0 for dock

+++ src/main/java/com/dji/sample/manage/model/receiver/WirelessLinkStateReceiver.java
@@ -15,7 +15,7 @@ public class WirelessLinkStateReceiver {
 
     private Integer downloadQuality;
 
-    private Integer frequencyBand;
+    private Double frequencyBand;
 
     private Integer upwardQuality;
